debugging new vie receive message

+++ lab2_cs7610/membership_protocol.cpp
@@ -613,7 +613,7 @@ void handle_messages(char* buf, uint32_t ty, fd_set tcp_writefds , int fdmax, ui
         }
         case 3:{
 
-            NEWVIEW_MESG *b = (NEWVIEW_MESG*) buf;
+            NEWVIEW_MESG* b = (NEWVIEW_MESG *) buf;
 
 
 
@@ -626,7 +626,7 @@ void handle_messages(char* buf, uint32_t ty, fd_set tcp_writefds , int fdmax, ui
             //}
             //cout<<"\n";
 
-            view_id = b->newview_id;
+            view_id = (uint32_t) b->newview_id;
             membership_list.assign( b->member_list , b->member_list+ b->no_members);
 
             //TODO: When a peer updates its view add the new members to the heartbeat timeout map and remove the
